// src/app/models/enums.model.ts
export enum OrigenMaterialLunar {
  Igneas = 'Ígneas',
  Metamoficas = 'Metamórficas',
  Sedimentarias = 'Sedimentarias',
}

export enum ClasificacionMaterialLunar {
  RocasConstruccion = 'Rocas de construcción: Son aquellas rocas que se usan para la construcción de diferentes estructuras.',
  RocaOrnamental = 'Roca ornamental: Son todas aquellas rocas de uso decorativo,integrado o no en edificio o estructuras superiores.',
  Utensilios = 'Utensilios: Rocas de uso en utensilios para el hombre.',
  PiedraMachada = 'Piedras machacadas: Áridos, ripios, agregados, etc. que son muy utilizados en la construcción como material de relleno.'
}

export enum TexturaMaterialLunar {
  Vitrea = 'Vítrea',
  Afanitica = 'Afanítica',
  Faneritica = 'Fanerítica',
}